backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.1
torch==2.3.1
transformers==4.41.2
sentencepiece==0.2.0
indic-transliteration==2.3.52
langdetect==1.0.9
faiss-cpu==1.12.0
gunicorn==23.0.0
